Update CrossVersionUtil.scala (#426)

+++ core/src/main/scala/sbt/internal/librarymanagement/cross/CrossVersionUtil.scala
@@ -64,10 +64,10 @@ object CrossVersionUtil {
    * Compatible versions include 2.10.0-1 and 2.10.1-M1 for Some(2, 10), but not 2.10.0-RC1.
    */
   private[sbt] def scalaApiVersion(v: String): Option[(Long, Long)] = v match {
-    case ReleaseV(x, y, _, _)                     => Some((x.toLong, y.toLong))
-    case BinCompatV(x, y, _, _, _)                => Some((x.toLong, y.toLong))
-    case NonReleaseV_1(x, y, z, _) if z.toInt > 0 => Some((x.toLong, y.toLong))
-    case _                                        => None
+    case ReleaseV(x, y, _, _)                      => Some((x.toLong, y.toLong))
+    case BinCompatV(x, y, _, _, _)                 => Some((x.toLong, y.toLong))
+    case NonReleaseV_1(x, y, z, _) if z.toLong > 0 => Some((x.toLong, y.toLong))
+    case _                                         => None
   }
 
   private[sbt] def partialVersion(s: String): Option[(Long, Long)] =
